refactor: 사용하지 않는 Auth 폴더 삭제
