+++ src/main/kotlin/gomushin/backend/auth/MainController.kt
@@ -1,4 +1,4 @@
-package gomushin.backend.auth.presentation
+package gomushin.backend.auth
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
