--- src/main/kotlin/gomushin/backend/auth/presentation/ApiPath.kt
@@ -1,4 +1,0 @@
-package gomushin.backend.auth.presentation
-
-object ApiPath {
-}
